refactor(tests): move _validate_splits test to dataset-level module
